+++ ts/masters/highcharts.src.ts
@@ -33,7 +33,7 @@ import Chart from '../Core/Chart/Chart.js';
 import ScrollablePlotArea from '../Extensions/ScrollablePlotArea.js';
 import StackingAxis from '../Core/Axis/Stacking/StackingAxis.js';
 import StackItem from '../Core/Axis/Stacking/StackItem.js';
-import DataTable from '../Data/DataTable.js';
+import DataTableBase from '../Data/DataTableBase.js';
 import Series from '../Core/Series/Series.js';
 import SeriesRegistry from '../Core/Series/SeriesRegistry.js';
 import '../Series/Line/LineSeries.js';
@@ -59,7 +59,7 @@ G.Axis = Axis;
 G.Chart = Chart;
 G.Color = Color;
 G.DataLabel = DataLabel;
-G.DataTable = DataTable;
+G.DataTable = DataTableBase;
 G.Fx = Fx;
 G.HTMLElement = HTMLElement;
 G.Legend = Legend;
